Pip 19.3+ improves handling of cache
Pip 19.3 and greater improves handling of the local cache and
communication with HTTP-based pip indexes. In particular, it
resolves a problem in which it provides a wrong header to a
pip index and triggers a 304 response, resulting in the client
wrongly concluding the package is available locally.

Resolve issue by ensuring that pip is upgraded before all other packages
are installed and by choosing the latest release.